Move cache-collection to $app-target/cache
A new collection "cache" is added to the project app where the cache is
stored independently of ediarum.web.
The cache file names now don't contain the app-target because it is the
same for one project.
IMPORTANT: the app-target must be the first parameter for all functions
using edwebapi:load-map-from-cache.