Add edweb relation functions and other functions
IMPROVEMENTS and NEW FUNCTIONS for HANDLING of RELATIONS

For all loaded relation items the key "is" describes if the item is
subject or object of the relation:

 - Update for all edweb:load-relations-*

It is possible to load relations not only for the current item but
also for an id:

 - Added edweb:load-relations-for-subject-with-id
 - Added edweb:load-relations-for-object-with-id
 - Added edweb:load-relations-for-subject-with-id-path
 - Added edweb:load-relations-for-object-with-id-path
 - Added edweb:load-relations-with-id

Relation of second degree can now be loaded. They are handled like
other relations. The predicate contains the relation item of the
first degree:

 - Added edweb:load-second-degree-relations

MORE NEW FUNCTIONS

 - Added edweb:add-detail-link Add a link with label
 - Added edweb:insert-count-of-groups Count objects after grouping
 - Added edweb:template-do-if-not-empty Process nodes only if xpath in current xml is not empty
 - Added edweb:template-for-each-group Equivalent to template:each but includes grouping