Update changes for eXist-db 5.3
In eXist-db 5.3 the functx-library is installed per default. To avoid
duplicates the functx-delcarations in ediarum.web must be removed:

- Change to functx module as import, remove functx declarations, and add
  functx dependency.
- Bugfix: update parameters for call of functx:change-element-ns-deep

In eXist 5 the storing of binary files doesn't work properly. There
exists an encoding problem, and special characters aren't saved
correctly. A workaround for the storing the json cache files must be
implemented:

- Wrap and store cache files as xml.